Consolidate test Span sources
