+++ src/Compilers/Test/Utilities/CSharp/CSharpTestBase.cs
@@ -2211,7 +2211,7 @@ namespace System.Diagnostics.CodeAnalysis
         protected static CSharpCompilation CreateCompilationWithSpan(CSharpTestSource tree, CSharpCompilationOptions options = null, CSharpParseOptions parseOptions = null)
         {
             var reference = CreateCompilation(
-                SpanSource,
+                TestSources.Span,
                 options: TestOptions.UnsafeReleaseDll);
 
             reference.VerifyDiagnostics();
@@ -2228,7 +2228,7 @@ namespace System.Diagnostics.CodeAnalysis
         protected static CSharpCompilation CreateCompilationWithMscorlibAndSpan(CSharpTestSource text, CSharpCompilationOptions options = null, CSharpParseOptions parseOptions = null)
         {
             var reference = CreateEmptyCompilation(
-                SpanSource,
+                TestSources.Span,
                 references: new List<MetadataReference>() { Net451.mscorlib, Net451.SystemCore, Net451.MicrosoftCSharp },
                 options: TestOptions.UnsafeReleaseDll);
 
@@ -2245,7 +2245,7 @@ namespace System.Diagnostics.CodeAnalysis
 
         protected static CSharpCompilation CreateCompilationWithMscorlibAndSpanSrc(string text, CSharpCompilationOptions options = null, CSharpParseOptions parseOptions = null)
         {
-            var textWitSpan = new string[] { text, SpanSource };
+            var textWitSpan = new string[] { text, TestSources.Span };
             var comp = CreateEmptyCompilation(
                 textWitSpan,
                 references: new List<MetadataReference>() { MscorlibRef_v4_0_30316_17626, SystemCoreRef, CSharpRef },
@@ -2254,289 +2254,6 @@ namespace System.Diagnostics.CodeAnalysis
 
             return comp;
         }
-
-        protected static readonly string SpanSource = @"
-namespace System
-    {
-        public readonly ref struct Span<T>
-        {
-            private readonly T[] arr;
-
-            public ref T this[int i] => ref arr[i];
-            public override int GetHashCode() => 1;
-            public int Length { get; }
-            public bool IsEmpty => Length == 0;
-
-            unsafe public Span(void* pointer, int length)
-            {
-                this.arr = Helpers.ToArray<T>(pointer, length);
-                this.Length = length;
-            }
-
-            public Span(T[] arr)
-            {
-                this.arr = arr;
-                this.Length = arr.Length;
-            }
-
-            public Span(T[] arr, int start, int length)
-            {
-                this.arr = new T[length];
-                Array.Copy(arr, start, this.arr, 0, length);
-                this.Length = length;
-            }
-
-            public void CopyTo(Span<T> other) { }
-
-            /// <summary>Gets an enumerator for this span.</summary>
-            public Enumerator GetEnumerator() => new Enumerator(this);
-
-            /// <summary>Enumerates the elements of a <see cref=""Span{T}""/>.</summary>
-            public ref struct Enumerator
-            {
-                /// <summary>The span being enumerated.</summary>
-                private readonly Span<T> _span;
-                /// <summary>The next index to yield.</summary>
-                private int _index;
-
-                /// <summary>Initialize the enumerator.</summary>
-                /// <param name=""span"">The span to enumerate.</param>
-                internal Enumerator(Span<T> span)
-                {
-                    _span = span;
-                    _index = -1;
-                }
-
-                /// <summary>Advances the enumerator to the next element of the span.</summary>
-                public bool MoveNext()
-                {
-                    int index = _index + 1;
-                    if (index < _span.Length)
-                    {
-                        _index = index;
-                        return true;
-                    }
-
-                    return false;
-                }
-
-                /// <summary>Gets the element at the current position of the enumerator.</summary>
-                public ref T Current
-                {
-                    get => ref _span[_index];
-                }
-            }
-
-            public static implicit operator Span<T>(T[] array) => new Span<T>(array);
-        }
-
-        public readonly ref struct ReadOnlySpan<T>
-        {
-            private readonly T[] arr;
-
-            public ref readonly T this[int i] => ref arr[i];
-            public override int GetHashCode() => 2;
-            public int Length { get; }
-            public bool IsEmpty => Length == 0;
-
-            unsafe public ReadOnlySpan(void* pointer, int length)
-            {
-                this.arr = Helpers.ToArray<T>(pointer, length);
-                this.Length = length;
-            }
-
-            public ReadOnlySpan(T[] arr)
-            {
-                this.arr = arr;
-                this.Length = arr.Length;
-            }
-
-            public ReadOnlySpan(T[] arr, int start, int length)
-            {
-                this.arr = new T[length];
-                Array.Copy(arr, start, this.arr, 0, length);
-                this.Length = length;
-            }
-
-            public void CopyTo(Span<T> other) { }
-
-            /// <summary>Gets an enumerator for this span.</summary>
-            public Enumerator GetEnumerator() => new Enumerator(this);
-
-            /// <summary>Enumerates the elements of a <see cref=""Span{T}""/>.</summary>
-            public ref struct Enumerator
-            {
-                /// <summary>The span being enumerated.</summary>
-                private readonly ReadOnlySpan<T> _span;
-                /// <summary>The next index to yield.</summary>
-                private int _index;
-
-                /// <summary>Initialize the enumerator.</summary>
-                /// <param name=""span"">The span to enumerate.</param>
-                internal Enumerator(ReadOnlySpan<T> span)
-                {
-                    _span = span;
-                    _index = -1;
-                }
-
-                /// <summary>Advances the enumerator to the next element of the span.</summary>
-                public bool MoveNext()
-                {
-                    int index = _index + 1;
-                    if (index < _span.Length)
-                    {
-                        _index = index;
-                        return true;
-                    }
-
-                    return false;
-                }
-
-                /// <summary>Gets the element at the current position of the enumerator.</summary>
-                public ref readonly T Current
-                {
-                    get => ref _span[_index];
-                }
-            }
-
-            public static implicit operator ReadOnlySpan<T>(T[] array) => array == null ? default : new ReadOnlySpan<T>(array);
-        }
-
-        public readonly ref struct SpanLike<T>
-        {
-            public readonly Span<T> field;
-        }
-
-        public enum Color: sbyte
-        {
-            Red,
-            Green,
-            Blue
-        }
-
-        public static unsafe class Helpers
-        {
-            public static T[] ToArray<T>(void* ptr, int count)
-            {
-                if (ptr == null)
-                {
-                    return null;
-                }
-
-                if (typeof(T) == typeof(sbyte))
-                {
-                    var arr = new sbyte[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((sbyte*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(byte))
-                {
-                    var arr = new byte[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((byte*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(short))
-                {
-                    var arr = new short[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((short*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(ushort))
-                {
-                    var arr = new ushort[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((ushort*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(int))
-                {
-                    var arr = new int[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((int*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(uint))
-                {
-                    var arr = new uint[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((uint*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(long))
-                {
-                    var arr = new long[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((long*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(ulong))
-                {
-                    var arr = new ulong[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((ulong*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(char))
-                {
-                    var arr = new char[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((char*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                if (typeof(T) == typeof(Color))
-                {
-                    var arr = new Color[count];
-                    for(int i = 0; i < count; i++)
-                    {
-                        arr[i] = ((Color*)ptr)[i];
-                    }
-
-                    return (T[])(object)arr;
-                }
-
-                throw new Exception(""add a case for: "" + typeof(T));
-            }
-        }
-    }";
         #endregion
 
         #region Index and Range
