+++ src/Compilers/Test/Utilities/CSharp/TestSources.cs
@@ -18,6 +18,7 @@ namespace System
         public ref T this[int i] => ref arr[i];
         public override int GetHashCode() => 1;
         public int Length { get; }
+        public bool IsEmpty => Length == 0;
 
         unsafe public Span(void* pointer, int length)
         {
@@ -29,6 +30,13 @@ namespace System
         {
             this.arr = arr;
             this.Length = arr is null ? 0 : arr.Length;
+        }
+
+        public Span(T[] arr, int start, int length)
+        {
+            this.arr = new T[length];
+            Array.Copy(arr, start, this.arr, 0, length);
+            this.Length = length;
         }
 
         public void CopyTo(Span<T> other) { }
@@ -76,12 +84,7 @@ namespace System
 
         public static implicit operator ReadOnlySpan<T>(Span<T> span) => new ReadOnlySpan<T>(span.arr);
 
-        public Span<T> Slice(int offset, int length)
-        {
-            var copy = new T[length];
-            Array.Copy(arr, offset, copy, 0, length);
-            return new Span<T>(copy);
-        }
+        public Span<T> Slice(int offset, int length) => new Span<T>(this.arr, offset, length);
     }
 
     public readonly ref struct ReadOnlySpan<T>
@@ -91,6 +94,7 @@ namespace System
         public ref readonly T this[int i] => ref arr[i];
         public override int GetHashCode() => 2;
         public int Length { get; }
+        public bool IsEmpty => Length == 0;
 
         unsafe public ReadOnlySpan(void* pointer, int length)
         {
@@ -102,6 +106,13 @@ namespace System
         {
             this.arr = arr;
             this.Length = arr is null ? 0 : arr.Length;
+        }
+
+        public ReadOnlySpan(T[] arr, int start, int length)
+        {
+            this.arr = new T[length];
+            Array.Copy(arr, start, this.arr, 0, length);
+            this.Length = length;
         }
 
         public void CopyTo(Span<T> other) { }
@@ -147,14 +158,10 @@ namespace System
 
         public static implicit operator ReadOnlySpan<T>(T[] array) => array == null ? default : new ReadOnlySpan<T>(array);
 
+        // NOTE: This is defined on String in the BCL (and the target type is non-generic ReadOnlySpan<char>).
         public static implicit operator ReadOnlySpan<T>(string stringValue) => string.IsNullOrEmpty(stringValue) ? default : new ReadOnlySpan<T>((T[])(object)stringValue.ToCharArray());
 
-        public ReadOnlySpan<T> Slice(int offset, int length)
-        {
-            var copy = new T[length];
-            Array.Copy(arr, offset, copy, 0, length);
-            return new ReadOnlySpan<T>(copy);
-        }
+        public ReadOnlySpan<T> Slice(int offset, int length) => new ReadOnlySpan<T>(this.arr, offset, length);
     }
 
     public readonly ref struct SpanLike<T>
@@ -178,12 +185,12 @@ namespace System
                 return null;
             }
 
-            if (typeof(T) == typeof(int))
+            if (typeof(T) == typeof(sbyte))
             {
-                var arr = new int[count];
+                var arr = new sbyte[count];
                 for(int i = 0; i < count; i++)
                 {
-                    arr[i] = ((int*)ptr)[i];
+                    arr[i] = ((sbyte*)ptr)[i];
                 }
 
                 return (T[])(object)arr;
@@ -195,6 +202,72 @@ namespace System
                 for(int i = 0; i < count; i++)
                 {
                     arr[i] = ((byte*)ptr)[i];
+                }
+
+                return (T[])(object)arr;
+            }
+
+            if (typeof(T) == typeof(short))
+            {
+                var arr = new short[count];
+                for(int i = 0; i < count; i++)
+                {
+                    arr[i] = ((short*)ptr)[i];
+                }
+
+                return (T[])(object)arr;
+            }
+
+            if (typeof(T) == typeof(ushort))
+            {
+                var arr = new ushort[count];
+                for(int i = 0; i < count; i++)
+                {
+                    arr[i] = ((ushort*)ptr)[i];
+                }
+
+                return (T[])(object)arr;
+            }
+
+            if (typeof(T) == typeof(int))
+            {
+                var arr = new int[count];
+                for(int i = 0; i < count; i++)
+                {
+                    arr[i] = ((int*)ptr)[i];
+                }
+
+                return (T[])(object)arr;
+            }
+
+            if (typeof(T) == typeof(uint))
+            {
+                var arr = new uint[count];
+                for(int i = 0; i < count; i++)
+                {
+                    arr[i] = ((uint*)ptr)[i];
+                }
+
+                return (T[])(object)arr;
+            }
+
+            if (typeof(T) == typeof(long))
+            {
+                var arr = new long[count];
+                for(int i = 0; i < count; i++)
+                {
+                    arr[i] = ((long*)ptr)[i];
+                }
+
+                return (T[])(object)arr;
+            }
+
+            if (typeof(T) == typeof(ulong))
+            {
+                var arr = new ulong[count];
+                for(int i = 0; i < count; i++)
+                {
+                    arr[i] = ((ulong*)ptr)[i];
                 }
 
                 return (T[])(object)arr;
